refactor: share frontmatter metadata parsing
